oopsy: grow live list after combat, make scrollable
Unfortunately, ffxiv eats mouse wheels and keyboard events when it
has the focus.  This is true even if "Enable clickthru" is not true
on your overlay.

It's possible to alt+tab to ACT and then mousewheel on the UI, but
this is nothing short of extremely awkward.

For now, just add a wide-ish scrollbar to make it reasonable to
click on.  Sorry.